Add Node.js time parsing tests
