+++ README.md
@@ -1,0 +1,14 @@
+# Enkeltimer
+
+## Running Tests
+
+This project includes a small Node.js test to validate time parsing.
+
+1. Ensure you have Node.js installed.
+2. From the project root, run:
+
+```bash
+node test/test.js
+```
+
+You should see `All tests passed.` if everything works correctly.
