+++ tasks.html
@@ -181,6 +181,7 @@
   
   <button id="refresh-button" class="icon-btn refresh-btn-fixed" title="Oppdater data">🔄</button>
   
+  <script src="time-utils.js"></script>
   <script src="script.js"></script>
   <script src="theme.js"></script> 
   <script src="gamification.js"></script>
